+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v3.0.0
+## v3.0.0 (2017-02-02)
 
 This release contains a huge facelift in Kodo. The DSL is now more simple but yet powerful.